pyproject.toml
[tool.poetry]
name = "waypath"
version = "0.7.0"
description = "Waypath is an AI-driven workflow automation platform focused initially on order processing"
authors = ["LukV <luk.verhelst@gmail.com>"]
readme = "README.md"
license = "LICENSE"
packages = [{ include = "core", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
fastapi = "^0.115.12"
uvicorn = "^0.34.2"
typer = "^0.15.2"
dotenv = "^0.9.9"
sqlalchemy = {extras = ["mypy"], version = "^2.0.40"}
python-jose = "^3.4.0"
ulid-py = "^1.1.0"
passlib = "^1.7.4"
pydantic = {extras = ["email"], version = "^2.11.3"}

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.7"
mypy = "^1.15.0"
pre-commit = "^4.2.0"
commitizen = "^4.6.0"

[tool.poetry.scripts]
wpath = "core.cli:app"

[tool.mypy]
plugins = ["sqlalchemy.ext.mypy.plugin", "pydantic.mypy"]

[tool.ruff]
# your ruff config here
line-length = 100

[tool.ruff.format]
quote-style = "preserve"
indent-style = "space"
line-ending = "auto"

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.commitizen] # https://commitizen-tools.github.io/commitizen/config/
bump_message = "bump: v$current_version → v$new_version"
tag_format = "v$version"
update_changelog_on_bump = true
version_provider = "poetry"
